Upgrade to Uno 5.1 & .NET 8
