--- Celbridge/Celbridge.Base/Celbridge.Base.csproj
@@ -1,12 +1,0 @@
-<Project Sdk="Microsoft.Build.NoTargets/3.7.0">
-	<PropertyGroup>
-		<!-- NOTE: The TargetFramework is required by MSBuild but not used as this project is not built. -->
-		<TargetFramework>net7.0</TargetFramework>
-		<EnableDefaultItems>false</EnableDefaultItems>
-	</PropertyGroup>
-
-	<ItemGroup>
-		<None Include="**\*" Exclude="obj\**;bin\**;*.csproj" />
-		<None Update="AppHead.xaml.cs" DependentUpon="AppHead.xaml" />
-	</ItemGroup>
-</Project>
