--- Celbridge/Celbridge.Base/base.props
@@ -1,29 +1,0 @@
-<Project>
-	<ItemGroup>
-		<PackageReference Include="Uno.Resizetizer" />
-	</ItemGroup>
-
-	<ItemGroup>
-		<None Include="$(MSBuildThisFileDirectory)AppHead.xaml" />
-		<ApplicationDefinition Include="$(MSBuildThisFileDirectory)AppHead.xaml"
-								SubType="Designer"
-								XamlRuntime="WinUI"
-								Generator="MSBuild:Compile"
-								Link="AppHead.xaml" />
-		<Compile Include="$(MSBuildThisFileDirectory)AppHead.xaml.cs"
-				XamlRuntime="WinUI"
-				DependentUpon="AppHead.xaml"
-				Link="AppHead.xaml.cs" />
-		<UnoIcon Include="$(MSBuildThisFileDirectory)Icons\logo.png"
-				ForegroundFile="$(MSBuildThisFileDirectory)Icons\logo.png"
-				Color="#00000000" />
-		<UnoSplashScreen
-			Include="$(MSBuildThisFileDirectory)Splash\splash_screen.svg"
-			BaseSize="128,128"
-			Color="#FFFFFF" />
-		<!-- NOTE: Files explicitly linked to display in the head projects for clarity. -->
-		<None Include="$(MSBuildThisFileDirectory)Icons\iconapp.svg" Link="Icons\iconapp.svg" />
-		<None Include="$(MSBuildThisFileDirectory)Icons\appconfig.svg" Link="Icons\appconfig.svg" />
-		<None Include="$(MSBuildThisFileDirectory)Splash\splash_screen.svg" Link="Splash\splash_screen.svg" />
-	</ItemGroup>
-</Project>
