+++ Celbridge/Celbridge.Legacy/Views/InspectorPanel.xaml
@@ -3,7 +3,7 @@
     xmlns="http://schemas.microsoft.com/winfx/2006/xaml/presentation"
     xmlns:x="http://schemas.microsoft.com/winfx/2006/xaml"
     xmlns:utils="using:Celbridge.Legacy.Utils"
-    xmlns:toolkit="using:CommunityToolkit.WinUI.UI.Controls"
+    xmlns:toolkit="using:CommunityToolkit.WinUI.Controls"
     xmlns:win="http://schemas.microsoft.com/winfx/2006/xaml/presentation" 
     xmlns:views="using:Celbridge.Legacy.Views"
     Background="{ThemeResource ApplicationPageBackgroundThemeBrush}"
@@ -89,20 +89,13 @@
                                   VerticalAlignment="Top"
                                   MinHeight="4"
                                   Height="4"
-                                  GripperCursor="Default"
                                   ResizeDirection="Auto"
                                   ResizeBehavior="BasedOnAlignment"
-                                  CursorBehavior="ChangeOnSplitterHover"
                                   Background="{StaticResource PanelBorderBrush}" 
                                   Foreground="{StaticResource PanelBackgroundBBrush}">
                 <toolkit:GridSplitter.RenderTransform>
                     <TranslateTransform Y="-2" />
                 </toolkit:GridSplitter.RenderTransform>
-                <toolkit:GridSplitter.Element>
-                    <Grid>
-                        <!-- No gripper -->
-                    </Grid>
-                </toolkit:GridSplitter.Element>
             </toolkit:GridSplitter>
         </Grid>
     </Grid>
