+++ Celbridge/Celbridge.MainApplication/Business/Models/AppConfig.cs
@@ -1,0 +1,6 @@
+namespace Celbridge.Business.Models;
+
+public record AppConfig
+{
+    public string? Environment { get; init; }
+}
