+++ Celbridge/Celbridge.MainApplication/GlobalUsings.cs
@@ -1,4 +1,4 @@
-﻿global using Celbridge.Legacy.Models;
+global using Celbridge.Legacy.Models;
 global using CommunityToolkit.Diagnostics;
 global using Microsoft.Extensions.DependencyInjection;
 global using Microsoft.Extensions.Hosting;
