+++ Celbridge/Celbridge.Shared/AppHead.xaml
@@ -1,0 +1,17 @@
+﻿<local:App x:Class="Celbridge.AppHead"
+       xmlns="http://schemas.microsoft.com/winfx/2006/xaml/presentation"
+       xmlns:x="http://schemas.microsoft.com/winfx/2006/xaml"
+       xmlns:wasm="http://platform.uno/wasm"
+       xmlns:local="using:Celbridge"
+       xmlns:mc="http://schemas.openxmlformats.org/markup-compatibility/2006"
+       mc:Ignorable="wasm">
+
+  <local:App.Resources>
+    <ResourceDictionary>
+      <ResourceDictionary.MergedDictionaries>
+        <ResourceDictionary Source="ms-appx:///Celbridge.CommonViews/Resources/AppResources.xaml" />
+      </ResourceDictionary.MergedDictionaries>
+    </ResourceDictionary>
+  </local:App.Resources>
+
+</local:App>
