+++ Celbridge/Celbridge.Shared/AppHead.xaml.cs
@@ -1,0 +1,27 @@
+using Microsoft.UI.Xaml;
+using Uno.Resizetizer;
+
+namespace Celbridge;
+public sealed partial class AppHead : App
+{
+    /// <summary>
+    /// Initializes the singleton application object. This is the first line of authored code
+    /// executed, and as such is the logical equivalent of main() or WinMain().
+    /// </summary>
+    public AppHead()
+    {
+        this.InitializeComponent();
+    }
+
+    /// <summary>
+    /// Invoked when the application is launched normally by the end user.  Other entry points
+    /// will be used such as when the application is launched to open a specific file.
+    /// </summary>
+    /// <param name="args">Details about the launch request and process.</param>
+    protected override void OnLaunched(LaunchActivatedEventArgs args)
+    {
+        base.OnLaunched(args);
+
+        MainWindow.SetWindowIcon();
+    }
+}
