+++ Celbridge/Celbridge.Shared/Celbridge.Shared.csproj
@@ -1,0 +1,12 @@
+<Project Sdk="Microsoft.Build.NoTargets">
+  <PropertyGroup>
+    <!-- NOTE: The TargetFramework is required by MSBuild but not used as this project is not built. -->
+    <TargetFramework>$(DotNetVersion)</TargetFramework>
+    <EnableDefaultItems>false</EnableDefaultItems>
+  </PropertyGroup>
+
+  <ItemGroup>
+    <None Include="**\*" Exclude="obj\**;bin\**;*.csproj" />
+    <None Update="AppHead.xaml.cs" DependentUpon="AppHead.xaml" />
+  </ItemGroup>
+</Project>
