+++ Celbridge/Celbridge.Skia.Gtk/Package.appxmanifest
@@ -1,43 +1,43 @@
 ﻿<?xml version="1.0" encoding="utf-8"?>
 
 <Package
-	xmlns="http://schemas.microsoft.com/appx/manifest/foundation/windows10"
-	xmlns:uap="http://schemas.microsoft.com/appx/manifest/uap/windows10"
-	xmlns:rescap="http://schemas.microsoft.com/appx/manifest/foundation/windows10/restrictedcapabilities"
-	IgnorableNamespaces="uap rescap">
+  xmlns="http://schemas.microsoft.com/appx/manifest/foundation/windows10"
+  xmlns:uap="http://schemas.microsoft.com/appx/manifest/uap/windows10"
+  xmlns:rescap="http://schemas.microsoft.com/appx/manifest/foundation/windows10/restrictedcapabilities"
+  IgnorableNamespaces="uap rescap">
 
-	<Identity
-		Name="Celbridge"
-		Publisher="O=Chris Gregan"
-		Version="1.0.0.0" />
+  <Identity
+    Name="Celbridge"
+    Publisher="O=An Tulcha"
+    Version="1.0.0.0" />
 
-	<Properties>
-		<DisplayName>Celbridge</DisplayName>
-		<PublisherDisplayName>Celbridge</PublisherDisplayName>
-	</Properties>
+  <Properties>
+    <DisplayName>Celbridge</DisplayName>
+    <PublisherDisplayName>Celbridge</PublisherDisplayName>
+  </Properties>
 
-	<Dependencies>
-		<TargetDeviceFamily Name="Windows.Universal" MinVersion="10.0.17763.0" MaxVersionTested="10.0.19041.0" />
-		<TargetDeviceFamily Name="Windows.Desktop" MinVersion="10.0.17763.0" MaxVersionTested="10.0.19041.0" />
-	</Dependencies>
+  <Dependencies>
+    <TargetDeviceFamily Name="Windows.Universal" MinVersion="10.0.17763.0" MaxVersionTested="10.0.19041.0" />
+    <TargetDeviceFamily Name="Windows.Desktop" MinVersion="10.0.17763.0" MaxVersionTested="10.0.19041.0" />
+  </Dependencies>
 
-	<Resources>
-		<Resource Language="x-generate"/>
-	</Resources>
+  <Resources>
+    <Resource Language="x-generate"/>
+  </Resources>
 
-	<Applications>
-		<Application Id="App"
-			Executable="$targetnametoken$.exe"
-			EntryPoint="$targetentrypoint$">
-			<uap:VisualElements
-				DisplayName="Celbridge"
-				Description="Celbridge">
-				<uap:SplashScreen />
-			</uap:VisualElements>
-		</Application>
-	</Applications>
+  <Applications>
+    <Application Id="App"
+      Executable="$targetnametoken$.exe"
+      EntryPoint="$targetentrypoint$">
+      <uap:VisualElements
+        DisplayName="Celbridge"
+        Description="Celbridge">
+        <uap:SplashScreen />
+      </uap:VisualElements>
+    </Application>
+  </Applications>
 
-	<Capabilities>
-		<rescap:Capability Name="runFullTrust" />
-	</Capabilities>
+  <Capabilities>
+    <rescap:Capability Name="runFullTrust" />
+  </Capabilities>
 </Package>
