--- Celbridge/Celbridge.Skia.Gtk/Properties/launchSettings.json
@@ -1,11 +1,0 @@
-{
-  "profiles": {
-    "Celbridge.Skia.Gtk": {
-      "commandName": "Project"
-    },
-    "WSL": {
-      "commandName": "WSL2",
-      "distributionName": ""
-    }
-  }
-}
