+++ Celbridge/Celbridge.Windows/Properties/PublishProfiles/win-arm64.pubxml
@@ -1,20 +1,22 @@
-<?xml version="1.0" encoding="utf-8"?>
+﻿<?xml version="1.0" encoding="utf-8"?>
 <!--
 https://go.microsoft.com/fwlink/?LinkID=208121.
 -->
 <Project ToolsVersion="4.0" xmlns="http://schemas.microsoft.com/developer/msbuild/2003">
 	<PropertyGroup>
 		<PublishProtocol>FileSystem</PublishProtocol>
-		<Platform>x64</Platform>
-		<RuntimeIdentifier>win10-x64</RuntimeIdentifier>
+		<Platform>arm64</Platform>
+		<RuntimeIdentifier>win-arm64</RuntimeIdentifier>
 		<PublishDir>bin\$(Configuration)\$(TargetFramework)\$(RuntimeIdentifier)\publish\</PublishDir>
 		<SelfContained>true</SelfContained>
 		<PublishSingleFile>False</PublishSingleFile>
 		<PublishReadyToRun Condition="'$(Configuration)' == 'Debug'">False</PublishReadyToRun>
 		<PublishReadyToRun Condition="'$(Configuration)' != 'Debug'">True</PublishReadyToRun>
-		<!--
-		See https://github.com/microsoft/CsWinRT/issues/373
-		<PublishTrimmed>True</PublishTrimmed>
+		<!-- Note: Trimming disabled by default as there may still be an issues with PublishTrimmed support: https://github.com/microsoft/CsWinRT/issues/373 -->
+		<!-- 
+		<PublishTrimmed Condition="'$(Configuration)' != 'Debug'">True</PublishTrimmed>
+		<TrimMode>partial</TrimMode>
+		<SuppressTrimAnalysisWarnings>true</SuppressTrimAnalysisWarnings> 
 		-->
 	</PropertyGroup>
 </Project>
