+++ Celbridge/Celbridge.Windows/Properties/PublishProfiles/win-x86.pubxml
@@ -6,15 +6,17 @@ https://go.microsoft.com/fwlink/?LinkID=208121.
 	<PropertyGroup>
 		<PublishProtocol>FileSystem</PublishProtocol>
 		<Platform>x86</Platform>
-		<RuntimeIdentifier>win10-x86</RuntimeIdentifier>
+		<RuntimeIdentifier>win-x86</RuntimeIdentifier>
 		<PublishDir>bin\$(Configuration)\$(TargetFramework)\$(RuntimeIdentifier)\publish\</PublishDir>
 		<SelfContained>true</SelfContained>
 		<PublishSingleFile>False</PublishSingleFile>
 		<PublishReadyToRun Condition="'$(Configuration)' == 'Debug'">False</PublishReadyToRun>
 		<PublishReadyToRun Condition="'$(Configuration)' != 'Debug'">True</PublishReadyToRun>
-		<!--
-		See https://github.com/microsoft/CsWinRT/issues/373
-		<PublishTrimmed>True</PublishTrimmed>
+		<!-- Note: Trimming disabled by default as there may still be an issues with PublishTrimmed support: https://github.com/microsoft/CsWinRT/issues/373 -->
+		<!-- 
+		<PublishTrimmed Condition="'$(Configuration)' != 'Debug'">True</PublishTrimmed>
+		<TrimMode>partial</TrimMode>
+		<SuppressTrimAnalysisWarnings>true</SuppressTrimAnalysisWarnings> 
 		-->
 	</PropertyGroup>
 </Project>
