--- Celbridge/Celbridge.Windows/web.config
@@ -1,44 +1,0 @@
-﻿<?xml version="1.0" encoding="UTF-8"?>
-<configuration>
-  <system.webServer>
-    <staticContent>
-      <remove fileExtension=".blat" />
-      <remove fileExtension=".dat" />
-      <remove fileExtension=".dll" />
-      <remove fileExtension=".webcil" />
-      <remove fileExtension=".json" />
-      <remove fileExtension=".wasm" />
-      <remove fileExtension=".woff" />
-      <remove fileExtension=".woff2" />
-      <mimeMap fileExtension=".blat" mimeType="application/octet-stream" />
-      <mimeMap fileExtension=".dll" mimeType="application/octet-stream" />
-      <mimeMap fileExtension=".webcil" mimeType="application/octet-stream" />
-      <mimeMap fileExtension=".dat" mimeType="application/octet-stream" />
-      <mimeMap fileExtension=".json" mimeType="application/json" />
-      <mimeMap fileExtension=".wasm" mimeType="application/wasm" />
-      <mimeMap fileExtension=".woff" mimeType="application/font-woff" />
-      <mimeMap fileExtension=".woff2" mimeType="application/font-woff" />
-    </staticContent>
-    <httpCompression>
-      <dynamicTypes>
-        <add mimeType="application/octet-stream" enabled="true" />
-        <add mimeType="application/wasm" enabled="true" />
-      </dynamicTypes>
-    </httpCompression>
-    <rewrite>
-      <rules>
-        <rule name="Serve subdir">
-          <match url=".*" />
-          <action type="Rewrite" url="wwwroot\{R:0}" />
-        </rule>
-        <rule name="SPA fallback routing" stopProcessing="true">
-          <match url=".*" />
-          <conditions logicalGrouping="MatchAll">
-            <add input="{REQUEST_FILENAME}" matchType="IsFile" negate="true" />
-          </conditions>
-          <action type="Rewrite" url="wwwroot\" />
-        </rule>
-      </rules>
-    </rewrite>
-  </system.webServer>
-</configuration>
