+++ Celbridge/Directory.Build.props
@@ -1,67 +1,20 @@
 <Project>
+  <PropertyGroup>
+    <DotNetVersion>net8.0</DotNetVersion>
+    <ImplicitUsings>enable</ImplicitUsings>
+    <Nullable>enable</Nullable>
 
-	<!--
-		If working on a single target framework, copy solution-config.props.sample to solution-config.props
-		and uncomment the appropriate lines in solution-config.props to build for the desired platforms only.
+    <ManagePackageVersionsCentrally>true</ManagePackageVersionsCentrally>
 
-		https://platform.uno/docs/articles/guides/solution-building-single-targetframework.html 
-	-->
-	<Import Project="solution-config.props" Condition="exists('solution-config.props')" />
-	
+    <!--
+      Adding NoWarn to remove build warnings
+      NU1507: Warning when there are multiple package sources when using CPM with no source mapping
+      NETSDK1201: Warning that specifying RID won't create self containing app
+      PRI257: Ignore default language (en) not being one of the included resources (eg en-us, en-uk)
+    -->
+    <NoWarn>$(NoWarn);NU1507;NETSDK1201;PRI257</NoWarn>
 
-	<PropertyGroup>
-		<ImplicitUsings>enable</ImplicitUsings>
-		<Nullable>enable</Nullable>
-
-		<DebugType>portable</DebugType>
-		<DebugSymbols>True</DebugSymbols>
-
-		<ManagePackageVersionsCentrally>true</ManagePackageVersionsCentrally>
-
-		<NoWarn>$(NoWarn);CA1416;NU1507</NoWarn>
-
-		<DefaultLanguage>en-us</DefaultLanguage>
-
-		<IsAndroid>false</IsAndroid>
-		<IsIOS>false</IsIOS>
-		<IsMac>false</IsMac>
-		<IsMacCatalyst>false</IsMacCatalyst>
-		<IsWinAppSdk>false</IsWinAppSdk>
-	</PropertyGroup>
-
-	<Choose>
-		<When Condition="$([MSBuild]::GetTargetPlatformIdentifier('$(TargetFramework)')) == 'android'">
-			<PropertyGroup>
-				<IsAndroid>true</IsAndroid>
-				<SupportedOSPlatformVersion>21.0</SupportedOSPlatformVersion>
-			</PropertyGroup>
-		</When>
-		<When Condition="$([MSBuild]::GetTargetPlatformIdentifier('$(TargetFramework)')) == 'ios'">
-			<PropertyGroup>
-				<IsIOS>true</IsIOS>
-				<SupportedOSPlatformVersion>14.2</SupportedOSPlatformVersion>
-			</PropertyGroup>
-		</When>
-		<When Condition="$([MSBuild]::GetTargetPlatformIdentifier('$(TargetFramework)')) == 'macos'">
-			<PropertyGroup>
-				<IsMac>true</IsMac>
-				<SupportedOSPlatformVersion>10.14</SupportedOSPlatformVersion>
-			</PropertyGroup>
-		</When>
-		<When Condition="$([MSBuild]::GetTargetPlatformIdentifier('$(TargetFramework)')) == 'maccatalyst'">
-			<PropertyGroup>
-				<IsMacCatalyst>true</IsMacCatalyst>
-				<SupportedOSPlatformVersion>14.0</SupportedOSPlatformVersion>
-			</PropertyGroup>
-		</When>
-		<When Condition="$(TargetFramework.Contains('windows10'))">
-			<PropertyGroup>
-				<IsWinAppSdk>true</IsWinAppSdk>
-				<SupportedOSPlatformVersion>10.0.18362.0</SupportedOSPlatformVersion>
-				<TargetPlatformMinVersion>10.0.18362.0</TargetPlatformMinVersion>
-				<RuntimeIdentifiers>win10-x86;win10-x64;win10-arm64</RuntimeIdentifiers>
-			</PropertyGroup>
-		</When>
-	</Choose>
-
+    <!-- Required for Hot Reload (See https://github.com/dotnet/sdk/issues/36666) -->
+    <IncludeSourceRevisionInInformationalVersion Condition="'$(Configuration)'=='Debug'">false</IncludeSourceRevisionInInformationalVersion>
+  </PropertyGroup>
 </Project>
