+++ Celbridge/Directory.Build.targets
@@ -1,6 +1,2 @@
-<Project>
-	<ItemGroup>
-		<!-- Removes native usings to avoid Ambiguous reference -->
-		<Using Remove="@(Using->HasMetadata('Platform'))" />
-	</ItemGroup>
+﻿<Project>
 </Project>
