+++ Celbridge/Directory.Packages.props
@@ -1,44 +1,44 @@
 <Project ToolsVersion="15.0">
   <ItemGroup>
     <PackageVersion Include="CommunityToolkit.Diagnostics" Version="8.2.2" />
-    <PackageVersion Include="CommunityToolkit.Mvvm" Version="8.2.1" />
-    <PackageVersion Include="CommunityToolkit.WinUI.UI.Controls" Version="7.1.2" />
+    <PackageVersion Include="CommunityToolkit.Mvvm" Version="8.2.2" />
+    <PackageVersion Include="CommunityToolkit.WinUI.Controls.Sizers" Version="8.0.240109" />
+    <PackageVersion Include="coverlet.collector" Version="6.0.1" />
+    <PackageVersion Include="FluentAssertions" Version="6.12.0" />
+    <PackageVersion Include="Microsoft.Extensions.DependencyInjection" Version="8.0.0" />
+    <PackageVersion Include="Microsoft.Extensions.Logging.Console" Version="8.0.0" />
+    <PackageVersion Include="Microsoft.NET.Test.Sdk" Version="17.9.0" />
+    <PackageVersion Include="Microsoft.Windows.SDK.BuildTools" Version="10.0.22621.3233" />
+    <PackageVersion Include="Microsoft.WindowsAppSDK" Version="1.5.240227000" />
+    <PackageVersion Include="Newtonsoft.Json" Version="13.0.3" />
+    <PackageVersion Include="NUnit" Version="4.1.0" />
+    <PackageVersion Include="NUnit3TestAdapter" Version="4.5.0" />
+    <PackageVersion Include="Serilog" Version="3.1.1" />
+    <PackageVersion Include="Serilog.Sinks.Debug" Version="2.0.0" />
+    <PackageVersion Include="SkiaSharp.Skottie" Version="2.88.7" />
+    <PackageVersion Include="SkiaSharp.Views.Uno.WinUI" Version="2.88.7" />
+    <PackageVersion Include="Uno.Core.Extensions.Logging.Singleton" Version="4.0.1" />
+    <PackageVersion Include="Uno.Extensions.Configuration" Version="4.0.0" />
+    <PackageVersion Include="Uno.Extensions.Core.WinUI" Version="4.0.0" />
+    <PackageVersion Include="Uno.Extensions.Hosting" Version="4.0.0" />
+    <PackageVersion Include="Uno.Extensions.Hosting.WinUI" Version="4.0.0" />
+    <PackageVersion Include="Uno.Extensions.Localization" Version="4.0.0" />
+    <PackageVersion Include="Uno.Extensions.Localization.WinUI" Version="4.0.0" />
+    <PackageVersion Include="Uno.Extensions.Logging.WinUI" Version="4.0.0" />
+    <PackageVersion Include="Uno.Resizetizer" Version="1.3.0" />
+    <PackageVersion Include="Uno.Toolkit.WinUI" Version="5.1.7" />
+    <PackageVersion Include="Uno.UI.Adapter.Microsoft.Extensions.Logging" Version="$(UnoVersion)" />
+    <PackageVersion Include="Uno.UniversalImageLoader" Version="1.9.36" />
+    <PackageVersion Include="Uno.WinUI" Version="$(UnoVersion)" />
+    <PackageVersion Include="Uno.WinUI.DevServer" Version="$(UnoVersion)" />
+    <PackageVersion Include="Uno.WinUI.Lottie" Version="$(UnoVersion)" />
+    <PackageVersion Include="Uno.WinUI.Skia.Gtk" Version="$(UnoVersion)" />
+    <!-- Legacy dependencies -->
     <PackageVersion Include="IronPython" Version="2.7.12" />
     <PackageVersion Include="IronPython.StdLib" Version="2.7.12" />
-    <PackageVersion Include="LiteDB" Version="5.0.17" />
-    <PackageVersion Include="Microsoft.CodeAnalysis.CSharp" Version="4.6.0" />
+    <PackageVersion Include="Microsoft.CodeAnalysis.CSharp" Version="4.9.2" />
     <PackageVersion Include="Microsoft.CodeDom.Providers.DotNetCompilerPlatform" Version="4.1.0" />
-    <PackageVersion Include="Microsoft.Extensions.Logging.Console" Version="7.0.0" />
-    <PackageVersion Include="coverlet.collector" Version="6.0.0" />
-    <PackageVersion Include="FluentAssertions" Version="6.11.0" />
-    <PackageVersion Include="Microsoft.NET.Test.Sdk" Version="17.7.0" />
-    <PackageVersion Include="Newtonsoft.Json" Version="13.0.3" />
-    <PackageVersion Include="NUnit" Version="3.13.3" />
-    <PackageVersion Include="NUnit3TestAdapter" Version="4.5.0" />
-    <PackageVersion Include="Microsoft.Extensions.DependencyInjection" Version="7.0.0" />
-    <PackageVersion Include="Microsoft.Windows.SDK.BuildTools" Version="10.0.22621.756" />
-    <PackageVersion Include="Microsoft.WindowsAppSDK" Version="1.3.230724000" />
-    <PackageVersion Include="OpenAI" Version="1.7.2" />
-    <PackageVersion Include="Scriban" Version="5.9.0" />
-    <PackageVersion Include="Serilog" Version="3.0.1" />
-    <PackageVersion Include="Serilog.Sinks.Debug" Version="2.0.0" />
-    <PackageVersion Include="SkiaSharp.Skottie" Version="2.88.6" />
-    <PackageVersion Include="SkiaSharp.Views.Uno.WinUI" Version="2.88.6" />
-    <PackageVersion Include="Uno.CommunityToolkit.WinUI.UI.Controls" Version="7.1.100" />
-    <PackageVersion Include="Uno.Core.Extensions.Logging.Singleton" Version="4.0.1" />
-    <PackageVersion Include="Uno.Extensions.Configuration" Version="2.5.11" />
-    <PackageVersion Include="Uno.Extensions.Hosting" Version="2.5.11" />
-    <PackageVersion Include="Uno.Extensions.Hosting.WinUI" Version="2.5.11" />
-    <PackageVersion Include="Uno.Extensions.Localization" Version="2.5.11" />
-    <PackageVersion Include="Uno.Extensions.Localization.WinUI" Version="2.5.11" />
-    <PackageVersion Include="Uno.Toolkit.WinUI" Version="4.2.22" />
-    <PackageVersion Include="Uno.Extensions.Toolkit.WinUI" Version="2.5.11" />
-    <PackageVersion Include="Uno.Resizetizer" Version="1.1.3" />
-    <PackageVersion Include="Uno.UI.Adapter.Microsoft.Extensions.Logging" Version="4.10.26" />
-    <PackageVersion Include="Uno.UniversalImageLoader" Version="1.9.36" />
-    <PackageVersion Include="Uno.WinUI" Version="4.10.26" />
-    <PackageVersion Include="Uno.WinUI.Lottie" Version="4.10.26" />
-    <PackageVersion Include="Uno.WinUI.RemoteControl" Version="4.10.26" />
-    <PackageVersion Include="Uno.WinUI.Skia.Gtk" Version="4.10.26" />
+    <PackageVersion Include="OpenAI" Version="1.10.0" />
+    <PackageVersion Include="Scriban" Version="5.9.1" />
   </ItemGroup>
 </Project>
