+++ Celbridge/global.json
@@ -1,0 +1,6 @@
+{
+  "msbuild-sdks": {
+    "Uno.Sdk": "5.1.87",
+    "Microsoft.Build.NoTargets": "3.7.56"
+  }
+}
