+++ Celbridge/nuget.config
@@ -1,0 +1,12 @@
+<?xml version="1.0" encoding="utf-8"?>
+<configuration>
+    <packageSources>
+        <!-- 
+            This specific line is required when building with .NET 8.0.100/VS 17.8 and earlier.
+            The error without this feed is as follows:
+
+            InvalidProjectFileException: The SDK 'Uno.Sdk' specified could not be found.
+         -->
+        <add key="NuGet official package source" value="https://api.nuget.org/v3/index.json" />
+    </packageSources>
+</configuration>
